# Conflicts:
#	native/cocos/3d/assets/Mesh.cpp
#	native/cocos/3d/assets/Mesh.h
#	native/cocos/3d/assets/Skeleton.cpp
#	native/cocos/3d/assets/Skeleton.h
#	native/cocos/3d/skeletal-animation/SkeletalAnimationUtils.cpp
#	native/cocos/3d/skeletal-animation/SkeletalAnimationUtils.h
#	native/cocos/bindings/manual/jsb_assets_manual.cpp
#	native/cocos/bindings/manual/jsb_conversions_spec.cpp
#	native/cocos/core/assets/Font.cpp
#	native/cocos/core/assets/Material.cpp
#	native/cocos/core/assets/Material.h
#	native/cocos/core/assets/RenderingSubMesh.h
#	native/cocos/core/assets/TextureBase.cpp
#	native/cocos/core/scene-graph/Node.cpp
#	native/cocos/renderer/frame-graph/ResourceAllocator.h
#	native/cocos/renderer/gfx-base/GFXDef.cpp
#	native/cocos/renderer/gfx-base/GFXInputAssembler.h
#	native/cocos/renderer/gfx-base/GFXRenderPass.h
#	native/cocos/renderer/gfx-base/GFXTexture.h
#	native/cocos/renderer/gfx-base/states/GFXGeneralBarrier.h
#	native/cocos/renderer/gfx-base/states/GFXSampler.h
#	native/cocos/renderer/gfx-base/states/GFXTextureBarrier.h
#	native/cocos/renderer/gfx-gles2/GLES2Commands.cpp
#	native/cocos/renderer/gfx-gles2/GLES2GPUObjects.h
#	native/cocos/renderer/gfx-gles3/GLES3Commands.cpp
#	native/cocos/renderer/gfx-gles3/GLES3GPUObjects.h
#	native/cocos/renderer/gfx-metal/MTLUtils.mm
#	native/cocos/renderer/pipeline/PipelineStateManager.cpp
#	native/cocos/renderer/pipeline/deferred/BloomStage.h
#	native/cocos/renderer/pipeline/shadow/ShadowFlow.cpp
#	native/cocos/scene/Pass.cpp
#	native/cocos/scene/Pass.h

+++ native/cocos/core/assets/TextureBase.h
@@ -206,7 +206,7 @@ public:
 
 protected:
     static gfx::Device *getGFXDevice();
-    static gfx::Format  getGFXPixelFormat(PixelFormat format);
+    static gfx::Format getGFXPixelFormat(PixelFormat format);
 
     gfx::Format getGFXFormat() const;
 
@@ -244,12 +244,12 @@ public:
     uint32_t _height{1};
 
 protected:
-    ccstd::string    _id;
+    ccstd::string _id;
     gfx::SamplerInfo _samplerInfo;
-    gfx::Sampler *   _gfxSampler{nullptr};
-    gfx::Device *    _gfxDevice{nullptr};
+    gfx::Sampler *_gfxSampler{nullptr};
+    gfx::Device *_gfxDevice{nullptr};
 
-    uint32_t _textureHash{0};
+    uint64_t _textureHash{0};
 
 private:
     CC_DISALLOW_COPY_MOVE_ASSIGN(TextureBase);
